refactor: move common serial implementation into its own class
